revert: validate_config_update
