+++ packages/rs-dpp/src/data_contract/config/methods/mod.rs
@@ -1,1 +1,1 @@
-mod validate_update;
+mod validate_config_update;
